Fixed font-size estimation
In one of the commits that added support for the mini, and further commits, a
new 'btnSize' parameter was introduced.

This however was not reflected in the algorithm that determines the size.

Without this fix, the text was always too wide (90px is more than the 72px my
stream deck has), and also no longer virtually centered.

This commit should address both these issues.

I've only tested this with the v2.